ironclad/requirements.txt
torch
torchvision
facenet-pytorch
pillow
faiss-cpu
Flask